fix: refactor support stakeholder handling
